+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+* Synchronized with Cobbletext 0.2.0
+  * Removed `Library.clearGlyphs()`, added `Engine.clearTiles()`.
+
 0.100.1 (2020-05-06)
 --------------------